Fixes SymbolLayer textOffset and iconOffset on iOS
There seem to be some styles that will be special cases to the rules used from the json spec
this PR adds the ability to add overrides, so we can be flexible in special cases. In this case
we were using a NSArray for textOffset and iconOffset when it should be a CGVector.