Fixed error when (Shape)source + layers is removed before they are added to the map
When the ShapeSource is loading images it is not yet added to the map.
However, React Native can decide to remove views and expects each component to return it’s child view via getChildAt()
This caused an error when getLayerAt() expected the layers/views to be part of mLayers but in reality were still part of mQueuedLayers.