autogenerate docs and template files precommit
This adds running autogenerate.js prior to any commit. If there are
any changes in the generated files, the autogenerate.js script exits
with a non-zero exit code, causing the commit to fail, and telling the
user to add those changed files to the commit before trying the commit
again.

I created this PR because I have noticed many commits where the
documentation hadn't been generated, then a later commit does generate
it which results in messy commits with unrelated changes. This should
ensure that any changes to generated files are committed with the actual
changes that caused them.