feat(location): pass smallestDisplacement as param
When a new LocationEngineRequest is build it is instantiated with
`displacement` set to `0` (default value).

Under this circumstance location updates are fired every second
(`.setFastestInterval` takes presedence).

This will cause location updates even when not moving.

Fortunatly `displacement` takes care of that.
Location updates will only be fired when the location has changed
according to the `displacement` value in meters.

In `MainApplication.java` add `RCTMGLPackage` either with or without a displacement value:
`packages.add(new RCTMGLPackage(5)); // 5 meters before location update`

* this has no iOS counterpart - no iOS dev :(
* how would this work when autolinking in the future?
* I'm no android dev, is this a good approach?