Update devDependencies
This also triggered some changes in our test setup because of react-native-testing-libraries
breaking changes.

I've reverted jests major update, as it was causing test issues,
which I think would be solved by updating react-native (which is
something I don't want to introduce in this PR).

Additionally I removed babel-jest as it's obsolete.