support dynamically changing local JSON in styleURL
When dynamically changing a `MapView`s `styleURL` prop that is a locally defined style.json `setReactStyleURL` will call `MapboxMap`s `setStyle`.  
Without a passed in Builder, it'll use the default one - which assumes that the style stings is an URI (which it is not). 

To prevent a `Mapbox error [HTTP] Unable to parse resourceUrl` error  we need to pass in our own Builder that uses `fromJSON` instead `fromUri`